backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.9.3",
    "fastapi>=0.128.7",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain-google-genai>=4.2.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.40.0",
]
